[JUnit5] Add support for optionally changing test directory (#25)
Initially, all projects were copied into a temporary directory. This is
useful in many cases where the directory needs to be pristinely copied.
However, it is also harder to debug things as the path is lightly
obfuscated.

This PR makes the feature optional. By default, the behaviour is to use
the project directory as the in-tree value and delete the `.gradle` and
`build` directories (i.e. `CLEAN_BUILD`). Orginal behaviour can be
configured by using `PRISTINE`.